Login with Chakra (ui done)
