+++ webapp/package.json
@@ -15,11 +15,14 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "framer-motion": "^10.0.2",
+    "mui": "^0.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
+    "react-remove-scroll": "^2.5.5",
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
+    "react-scroll": "^1.8.9",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
